requirements.txt
absl-py==0.13.0
altair==4.1.0
argon2-cffi==20.1.0
asgiref==3.4.1
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
base58==2.1.0
bleach==4.0.0
blinker==1.4
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
click==7.1.2
colorama==0.4.4
cycler==0.10.0
Cython==0.29.21
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
docx2pdf==0.1.7
entrypoints==0.3
fastapi==0.67.0
filelock==3.0.12
flatbuffers==1.12
gast==0.3.3
gensim==4.0.1
gitdb==4.0.7
GitPython==3.1.18
google-auth==1.34.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
grpcio==1.32.0
h11==0.12.0
h5py==2.10.0
huggingface-hub==0.0.12
idna==3.2
img2pdf==0.4.1
ipykernel==6.0.3
ipython==7.26.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keras-tcn==3.4.0
kiwisolver==1.3.1
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.6.2
notebook==6.4.2
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.1
pandocfilters==1.4.3
parso==0.8.2
pdfminer==20191125
pickleshare==0.7.5
pikepdf==2.15.0
Pillow==8.3.1
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.3
psutil==5.8.0
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
pydantic==1.8.2
pydeck==0.6.2
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2021.1
pywin32==227
pywinpty==1.1.3
PyYAML==5.4.1
pyzmq==22.2.1
regex==2021.7.6
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.7.0
seaborn==0.11.1
Send2Trash==1.7.1
sentence-transformers==2.0.0
sentencepiece==0.1.96
six==1.15.0
sklearn==0.0
smart-open==5.1.0
smmap==4.0.0
starlette==0.14.2
streamlit==0.86.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.0
tensorflow-addons==0.13.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
threadpoolctl==2.2.0
tokenizers==0.10.3
toml==0.10.2
toolz==0.11.1
torch==1.9.0+cu111
torchaudio==0.9.0
torchvision==0.10.0+cu111
tornado==6.1
tqdm==4.61.2
traitlets==5.0.5
transformers==4.9.1
typeguard==2.12.1
typing-extensions==3.7.4.3
tzlocal==2.1
Unidecode==1.2.0
urllib3==1.26.6
uvicorn==0.14.0
validators==0.18.2
watchdog==2.1.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
